The "Sorry for no Light Structs" Update
Introduced in this update is the Disk object. It's like a cooler plane. But like a sphere. A slice of one.

Also some optimization regarding reflection materials. 

Yeah, sorry for no light structs. My brain slowly is giving up following the last attempt at making multiple types of lights, and a box object. It's been harder and harder to find a lot of material that I can understand. The 15 year old brain of mine can only cheat around for so long around math until I hit the wall of great equations. Yeah, the ones with the greek letters and everything.

But I love this thing. It's stupid, but I want to make it into something that could actually help me create fun content. Maybe sometime soon my brain could click and finally become PROGRAMMING LORD OF RAYTRACING UNEVEN.

Computers are weird. But so am I. And as I promised to my grandma, I'll do what I love doing. Using these things to do funny stuff and make people happy. So, be happy. It is one heck of a good thing to feel.